rename *_definitions.py to  fields_definitions.py
