+++ ckanext-basedosdados/ckanext/basedosdados/validator/packages/__init__.py
@@ -1,2 +1,2 @@
 from ckanext.basedosdados.validator.packages.ckan_default import package
-from ckanext.basedosdados.validator.packages.dataset import dataset, dataset_definitions
+from ckanext.basedosdados.validator.packages.dataset import dataset, fields_definitions
