+++ ckanext-basedosdados/ckanext/basedosdados/validator/packages/dataset/dataset.py
@@ -10,7 +10,7 @@ from pydantic import (
     root_validator,
 )
 
-from ckanext.basedosdados.validator.packages.dataset.dataset_definitions import *
+from ckanext.basedosdados.validator.packages.dataset.fields_definitions import *
 from ckanext.basedosdados.validator.packages.ckan_default.package import _CkanDefaults
