+++ ckanext-basedosdados/ckanext/basedosdados/validator/resources/bdm/table/fields_definitions.py
